UI Hotfix
This is an urgent commit to fix the UI from the previous commit (check #7 ). It also cleans up the UI, adding a health bar UI slider to prep for the Damage System ( #3 ), removing the `Respawn` button, and anchoring the UI to a new panel. New fonts are included into TextMeshPro via the font asset generator.