Migration to new Input System (#16)
* Fixed Sliding

I made a very simple tweak to the Box Collider on all the objects and the player itself- creating a new Physic Material and setting the Dynamic and Static friction to 0, effectively removing the sliding issue altogether. (Check #10 )

* Setting up input system

I've more or less finished setting up the input system in the inspector and made the scripts compatible with the package

* First draft

I've rewritten most of the code again to implement the new input system ( #8 ). However, there are several bugs with this commit and it is very unstable.

* Bug fixes

I've fixed many bugs from my last commit, namely the lack of responsiveness

* Fixed Jitter

I've fixed the strange jitter coming from the `FixedUpdate()` method by using the RigidBody to handle movement rather than overriding the physics engine with the `transform.Translate()` method.

* Cleaning up

I've cleaned up my scripts to make the code more intuitive to edit. The new input system is more or less complete. I'm going to start a pull request and start the review process.